requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
exceptiongroup==1.3.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
